规范代码，某些变量增加初始化操作。

+++ src/ledger/ledger_manager.cpp
@@ -599,13 +599,13 @@ namespace bubi {
 		WebSocketServer::Instance().BroadcastMsg(protocol::CHAIN_LEDGER_HEADER, tmp_lcl_header.SerializeAsString());
 
 		// notice applied
-		for (int i = 0; i < closing_ledger_->apply_tx_frms_.size(); i++) {
+		for (size_t i = 0; i < closing_ledger_->apply_tx_frms_.size(); i++) {
 			TransactionFrm::pointer tx = closing_ledger_->apply_tx_frms_[i];
 			WebSocketServer::Instance().BroadcastChainTxMsg(tx->GetContentHash(), tx->GetSourceAddress(),
 				tx->GetResult(), tx->GetResult().code() == protocol::ERRCODE_SUCCESS ? protocol::ChainTxStatus_TxStatus_COMPLETE : protocol::ChainTxStatus_TxStatus_FAILURE);
 		}
 		// notice dropped
-		for (int i = 0; i < closing_ledger_->dropped_tx_frms_.size(); i++) {
+		for (size_t i = 0; i < closing_ledger_->dropped_tx_frms_.size(); i++) {
 			TransactionFrm::pointer tx = closing_ledger_->dropped_tx_frms_[i];
 			WebSocketServer::Instance().BroadcastChainTxMsg(tx->GetContentHash(), tx->GetSourceAddress(),
 				tx->GetResult(), tx->GetResult().code() == protocol::ERRCODE_SUCCESS ? protocol::ChainTxStatus_TxStatus_COMPLETE : protocol::ChainTxStatus_TxStatus_FAILURE);
